Move examples around
